Add support for mini-screens.
Let the screen constructor take an optional div as a parameter to build itself in.  This lets you put a screen in a specific element on a page without destroying the entire document body.  Anything inside this element will be destroy, and it defaults to document body just like the old system.