Add Content to the Frontend.
Move the random text into a proper Content class and hook it up to the field with lots of events.

Content now renders (after 1 second mock loading period) into a field to be displayed for a random duration between 0-10 seconds.  After the duration is up a new piece of content is loaded and the process repeats forever.  This is all still using the sample text content.