Cache the template preview, closes #58.
Wrapping the template preview code in a stale? block to only render and return the image if the etag doesn't match or there has been a modification.

Templates now can find the last_modified time of not just the template but also it's positions and background too in the event those update.