Remove an old pair of migrations.
This should have no effect on the final state of the migration, as
remove_players reverses the changes made by create_players. A slight
risk is that if anyone has a working copy corresponding to the short
time between when create_players and remove_players were added, they
will have a migration error.