Plugins: basic initialization framework
This allows plugins to request to be mounted as a rack app for routing,
as well as set up their own options for Concerto Configuration.

The next step will be to implement a system for controller and UI hooks
into the main application. Also, we're still a ways away from one-click
installation, so some work will be required there too.