Prevent unauthorized modification of send_errors
Security issue on the Error Reporting option on first-user registration.

Just because a field doesn't appear in the form doesn't mean an attacker
can't submit it. In this case I was able to modify the user registration
form (in a running system where the admin was alread registered), and
change the value without gaining admin rights.