Make auth! capable of handling paginated relations
The natural form for paginating indices with Kaminari is
`Content.all.page(1).per(20)`. When we call our auth! method to
remove inaccessible items, we lose that pagination metadata,
causing the page decoration to break. Further, even if we fixed it,
we'd have been left with the possibility of users seeing scarce or
empty pages if they lacked permissions to large numbers of items.

This solution currently requirees fetching the whole dataset, but
provides the user with the correct offset and number of entries,
without any change to the controller code. Test case is the Pages
controller which is now working again.