Make frontend fields stateless to resolve repeating content issues.
This eliminates the concept of a store for timeline data, which was just
going to the client anyway when cookie store was enabled. Instead, the
entire timeline worth of content will be sent to the client. This
timeline is not reliant on getting a cookie back which could have been
overwritten in the client by a simultaneous request.

If a reliable store for this data is found, the store concept could be
reinstated. However the only reason for this would seem to be if the
client could not store the whole timeline, which seems unlikely.

See #1329.